moved index functions to seperate file
