+++ .env
@@ -1,8 +1,10 @@
 # Algolia App Id
-VITE_APP_ID=""
+VITE_ALGOLIA_APP_ID=""
 # Algolia Index Name
-VITE_INDEX_NAME=""
+VITE_ALGOLIA_INDEX_NAME=""
 # Algolia Admin API Key
-VITE_ADMIN_KEY=""
+VITE_ALGOLIA_ADMIN_KEY=""
 # Algolia Search API Key
-VITE_SEARCH_KEY=""
+VITE_ALGOLIA_SEARCH_KEY=""
+# Algolia Url Base
+VITE_ALGOLIA_URL_BASE=""
